refactor: separate V1/V3 typography into dedicated directories
- Move Font.swift → V1/BezierFont.swift
- Extract V1 typography extensions → V1/BezierFont+Typography.swift
- Extract V3 typography extensions → V3/BTSemanticToken+Typography.swift
- Remove mixed Typography.swift

Directory structure:
  Typography/
  ├── V1/
  │   ├── BezierFont.swift
  │   └── BezierFont+Typography.swift
  ├── V3/
  │   ├── BTGlobalToken.swift
  │   ├── BTFontWeight.swift
  │   ├── BTSemanticToken.swift
  │   ├── BTSemanticToken+Style.swift
  │   └── BTSemanticToken+Typography.swift
  └── README.md
